\name{plot-methods}

\docType{methods}

\alias{plot-methods}
\alias{plot,ANY-method}
\alias{plot,specL-methods}
\alias{plot,specLSet-methods}

\title{Method for Function \code{plot} in Package \pkg{specL}}
\description{
    This methode has no additional arguments.
}

\value{
    The method plots on the current device.
}

\section{Methods}{
    \describe{
        \item{\code{signature(x = "specL")}}{
            Plots the specL determined ions.
        }

        \item{\code{signature(x = "specLSet")}}{
            Plots retention time versus retention time.
        }
    }
}
\keyword{methods}
%\keyword{ ~~ other possible keyword(s) ~~ }
