\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,ANY-method}
\title{Methods for Function \code{show} in Package \pkg{specL} ~~}
\description{
Methods for function \code{show} in package \pkg{specL} ~~
writes specL or specLSet objects to a file or console.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "specL")}}{
    Prints specL object data to the console.
}
\item{\code{signature(x = "specLSet")}}{
    Prints specL object data to the console.
}
}}
\keyword{methods}
